# [ROCm] Apply FP8 weights padding to values not divisible by 512 bytes on ROCm