Performance improvement for ROCm working around the hardware limitation.
> In GEMM, you can have significant Tagram channel hotspot problems on MI300 if the stride of a matrix is a multiple of 512 bytes in GEMM. This is especially true for TN transpose cases, which might increase the latency of VMEM instructions, resulting in a significant drop in performance. If it's possible (or makes sense), stride padding can be used to avoid any stride multiple of 512 bytes (for example, for TN F16 GEMM, lda = M + 128 when M%256==0) from the application when allocating memory for the matrices.

One requirement for this is for w8a8_block_fp8_matmul to support the non-contiguous weights, which it seems to already do, so the leftover assertion is obsolete.
While maintaining the same correctness, this shows the following latency improvement on ROCm:
amd/Llama-3.1-8B-Instruct-FP8-KV bs=64 in=512 out=512 tp=1:
5.95s -> 5.7s (4%)
amd/Llama-3.1-70B-Instruct-FP8-KV bs=64 4in=512 out=512 tp=1:
25.6s -> 24.3s (5%)
deepseek-ai/DeepSeek-R1 bs=64 in=256 out=256 tp=8:
26.1s -> 24.9 (5%)